14commit

+++ src/main/kotlin/org/ldv/appliwickle/controllers/AdminControllers/AdminProduitController.kt
@@ -71,7 +71,7 @@ class AdminProduitController(private val produitDAO: ProduitDAO,
         redirectAttributes: RedirectAttributes
     ): String {
         try {
-            val categorie = categorieDAO.findById(categorieId).orElseThrow {
+            val categorie = categorieDAO.findById(categorieId.toInt()).orElseThrow {
                 IllegalArgumentException("Catégorie introuvable")
             }
